Update local Nonce in tx injector
